backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic[email]==2.10.4
pywebpush==2.0.1
